==> customer-support-messages/00-prepare-dataset.ipynb <==
# Prepare the dataset

We load a dataset with messages from customers that are assigned to different categories (invoices, orders, etc.)

We will use this dataset to tune a prompt that can classify a message into one of these categories, with an accuracy as high as possible.

https://huggingface.co/datasets/bitext/Bitext-customer-support-llm-chatbot-training-dataset

pip install datasets --quiet

from datasets import load_dataset

dataset = load_dataset(
    "bitext/Bitext-customer-support-llm-chatbot-training-dataset", split="train"
)
dataset

CLASSES = list(set(dataset["category"]))
CLASSES

### CONSTANTS

N = 200  # Number of samples per class / category
TRAIN_TEST_SPLIT = 0.75  # split dataset in trian and test
TEST_VAL_SPLIT = 0.7  # spit test in test and validation

import pandas as pd
from datasets import Dataset

# Convert the dataset to a pandas DataFrame
df = pd.DataFrame(dataset)
# add an id column
df = df.reset_index(drop=False).rename(columns={"index": "id"})
# Group by the label column
grouped = df.groupby("category")
# Sample 100 records from each label
sampled_df = grouped.apply(lambda x: x.sample(n=N, random_state=42)).reset_index(
    drop=True
)
# Convert the sampled DataFrame back to a Hugging Face dataset
sampled_dataset = Dataset.from_pandas(sampled_df)
print(sampled_df["category"].value_counts())
shuffled_dataset = sampled_dataset.shuffle()

Select only the relevant columns and rename them according to the class ClassificationDataClass

selected_cols_dataset = shuffled_dataset.select_columns(
    ["category", "instruction", "id"]
)
renamed_dataset = selected_cols_dataset.rename_column(
    "category", "class_name"
).rename_column("instruction", "question")
full_dataset = renamed_dataset
full_dataset.to_csv("full_dataset.csv")
full_dataset

# Split the dataset into train and test (70% train, 30% test)
train_test_split = full_dataset.train_test_split(test_size=TRAIN_TEST_SPLIT)
# Further split the test set into validation and test sets (50% validation, 50% test)
val_test_split = train_test_split["test"].train_test_split(test_size=TEST_VAL_SPLIT)

train = train_test_split["train"]
train.to_csv("train.csv")
val = val_test_split["train"]
val.to_csv("val.csv")
test = val_test_split["test"]
test.to_csv("test.csv")

train.to_pandas()["class_name"].value_counts()

val.to_pandas()["class_name"].value_counts()

test.to_pandas()["class_name"].value_counts()
